Fixed rating function

+++ header/movies.h
@@ -6,18 +6,17 @@ class Movies{
 
         string title; //title of the movie
         string genre; //genre of the movie
-        int rating; // rating of the movie out of 10
+        double rating; // rating of the movie out of 10
         int releaseDate; //the year the movie was released
         int views; // amount of viewership
 
 public:
         Movies(string title, int rating, string genre, int releaseDate, int views); 
-        int getRating();
+        double getRating();
         string getTitle();
         string Genre();
         int getReleaseDate();
         int getViews();
-
 
 }
